backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cloudinary==1.42.2
dj-database-url==2.3.0
Django==5.1.6
djangorestframework==3.15.2
psycopg2==2.9.10
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
